tabletserver endtoend: new framework package
The utility functions were growing. I've moved them to a separate
framework package.
