+++ go/vt/tabletserver/endtoend/compatibility_test.go
@@ -11,10 +11,11 @@ import (
 	"github.com/youtube/vitess/go/mysql"
 	mproto "github.com/youtube/vitess/go/mysql/proto"
 	"github.com/youtube/vitess/go/sqltypes"
+	"github.com/youtube/vitess/go/vt/tabletserver/endtoend/framework"
 )
 
 func TestCharaterSet(t *testing.T) {
-	qr, err := newClient(defaultServer).Execute("select * from vtocc_test where intval=1", nil)
+	qr, err := framework.NewDefaultClient().Execute("select * from vtocc_test where intval=1", nil)
 	if err != nil {
 		t.Error(err)
 		return
@@ -55,7 +56,7 @@ func TestCharaterSet(t *testing.T) {
 }
 
 func TestInts(t *testing.T) {
-	client := newClient(defaultServer)
+	client := framework.NewDefaultClient()
 	defer client.Execute("delete from vtocc_ints", nil)
 
 	_, err := client.Execute(
@@ -155,7 +156,7 @@ func TestInts(t *testing.T) {
 }
 
 func TestFractionals(t *testing.T) {
-	client := newClient(defaultServer)
+	client := framework.NewDefaultClient()
 	defer client.Execute("delete from vtocc_fracts", nil)
 
 	_, err := client.Execute(
@@ -218,7 +219,7 @@ func TestFractionals(t *testing.T) {
 }
 
 func TestStrings(t *testing.T) {
-	client := newClient(defaultServer)
+	client := framework.NewDefaultClient()
 	defer client.Execute("delete from vtocc_strings", nil)
 
 	_, err := client.Execute(
@@ -312,7 +313,7 @@ func TestStrings(t *testing.T) {
 }
 
 func TestMiscTypes(t *testing.T) {
-	client := newClient(defaultServer)
+	client := framework.NewDefaultClient()
 	defer client.Execute("delete from vtocc_misc", nil)
 
 	_, err := client.Execute(
@@ -375,7 +376,7 @@ func TestMiscTypes(t *testing.T) {
 }
 
 func TestNull(t *testing.T) {
-	client := newClient(defaultServer)
+	client := framework.NewDefaultClient()
 	qr, err := client.Execute("select null from dual", nil)
 	if err != nil {
 		t.Error(err)
